requirements.txt

alembic==0.9.5
autopep8==1.5.7
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==6.7
dnspython==2.2.0
dominate==2.3.1
email-validator==1.1.3
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
markdown2==2.4.1
MarkupSafe==1.1.0
Pillow==9.0.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.21
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==5.0.2
six==1.10.0
SQLAlchemy==1.1.13
text-unidecode==1.3
toml==0.10.2
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
zipp==3.7.0